Force-reset the SPI block when you reset the host.
This is because the host might be mid-SPI-byte, and we don't want to leave any bits in the receive register. The SPI block itself won't let you clear them, so we just reset it.

Sadly the Rcc driver in the HAL has no useful methods, so we ensure we have the object (for exclusive access) then do some raw pointer writes.